# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.BaseDTO:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  dto.Menu:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      priority:
        type: integer
      target:
        type: string
      team:
        type: string
      url:
        type: string
    type: object
...
  param.Menu:
    properties:
      URL:
        maxLength: 1023
        minLength: 1
        type: string
      icon:
        maxLength: 50
        type: string
      id:
        type: integer
      name:
        maxLength: 50
        minLength: 1
        type: string
      parentId:
        minimum: 0
        type: integer
      priority:
        minimum: 0
        type: integer
      target:
        maxLength: 50
        type: string
      team:
        maxLength: 255
        type: string
    type: object
...
  vo.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/vo.Menu'
        type: array
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      priority:
        type: integer
      target:
        type: string
      team:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    AirPress 博客系统后端 API 文档,通过 swag 注释驱动自动生成。
    包含后台管理(/api/admin)与前台内容(/api/content)两组接口。
  title: AirPress API
  version: "1.0"
paths:
...
  /admin/menus:
    get:
      consumes:
      - application/json
      description: 返回所有菜单列表,支持按 team 与 priority 排序
      parameters:
      - collectionFormat: multi
        description: 排序字段,如 team,desc
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Menu'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 查询所有菜单
      tags:
      - Admin.Menu
    post:
      consumes:
      - application/json
      description: 创建一个新菜单,返回创建后的详情
      parameters:
      - description: 菜单参数
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/param.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 创建菜单
      tags:
      - Admin.Menu
  /admin/menus/{id}:
    delete:
      description: 根据菜单ID删除指定菜单
      parameters:
      - description: 菜单ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 删除菜单
      tags:
      - Admin.Menu
    get:
      description: 返回指定 ID 的菜单详情
      parameters:
      - description: 菜单ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 根据ID获取菜单
      tags:
      - Admin.Menu
    put:
      consumes:
      - application/json
      description: 根据菜单ID更新菜单信息
      parameters:
      - description: 菜单ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单参数
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/param.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto.Menu'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 更新菜单
      tags:
      - Admin.Menu
  /admin/menus/batch:
    delete:
      consumes:
      - application/json
      description: 根据菜单ID列表批量删除菜单
      parameters:
      - description: 菜单ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 批量删除菜单
      tags:
      - Admin.Menu
    post:
      consumes:
      - application/json
      description: 根据菜单参数列表批量创建菜单
      parameters:
      - description: 菜单参数列表
        in: body
        name: menus
        required: true
        schema:
          items:
            $ref: '#/definitions/param.Menu'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Menu'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 批量创建菜单
      tags:
      - Admin.Menu
    put:
      consumes:
      - application/json
      description: 根据菜单参数列表批量更新菜单
      parameters:
      - description: 菜单参数列表
        in: body
        name: menus
        required: true
        schema:
          items:
            $ref: '#/definitions/param.Menu'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Menu'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 批量更新菜单
      tags:
      - Admin.Menu
  /admin/menus/team/tree_view:
    get:
      consumes:
      - application/json
      description: 根据 team 查询参数返回菜单的树形结构,team 为空时返回全部
      parameters:
      - collectionFormat: multi
        description: 排序字段,如 priority,asc
        in: query
        items:
          type: string
        name: sort
        type: array
      - description: 菜单分组名称
        in: query
        name: team
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/vo.Menu'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 按分组返回树形菜单
      tags:
      - Admin.Menu
  /admin/menus/teams:
    get:
      description: 返回所有菜单的分组(team)列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 查询菜单分组
      tags:
      - Admin.Menu
  /admin/menus/tree_view:
    get:
      consumes:
      - application/json
      description: 返回所有菜单的树形结构
      parameters:
      - collectionFormat: multi
        description: 排序字段,如 team,desc
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/vo.Menu'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 树形菜单列表
      tags:
      - Admin.Menu
...
securityDefinitions:
  AdminApiKey:
    in: header
    name: Admin-Authorization
    type: apiKey
  ContentApiKey:
    in: header
    name: Content-Authorization
    type: apiKey
swagger: "2.0"